Updated login script file, and fixed script import
