+++ cryptono/src/pages/login/login.html
@@ -48,6 +48,6 @@
             <p class="security-note">🔐 Encrypted & Secure</p>
         </div>
     </div>
-    <script type="module" src="popup.js"></script>
+    <script type="module" src="login.js"></script>
 </body>
 </html>
